Fix non-GNU grep installer issue
The grep flag -P only exists on GNU grep, which is not the default grep
on macOS; the awk script used instead works and obtains the same value.

Also did some work on making it so that at least the amd64 version will
install on Apple Silicon Macs. This is suboptimal, but it should do the
job.